Add plymouth-poweroff.sh script
According to the systemd documentation "Immediately before executing the
actual system halt/poweroff/reboot/kexec systemd-shutdown will run all
executables in /usr/lib/systemd/system-shutdown/ and pass one arguments
to them: either "halt", "poweroff", "reboot" or "kexec", depending on
the chosen action. All executables in this directory are executed in
parallel, and execution of the action is not continued before all
executables finished."

In this directory we put the plymouth-poweroff.sh script to be executed
when powering off to show a black screen. This is actually shown only
after the rootfs has been already remounted in read-only so that we can
safely remove the power supply.

Signed-off-by: Carlo Caione <carlo@endlessm.com>