drm: Set a linear ramp for CLUTs before displaying the splash screen
When calling drmModeSetCrtc there's no guarantee that the CLUT contains
suitable values  (e.g. the hardware could currently be programmed for a
different colour bit depth). On some (AMD) hardware this is causing the
colors of the reboot / poweroff splash to appear scrambled.

To avoid this we initialize a suitable (linear) gamma ramp and set that
with drmModeCrtcSetGamma before displaying the splash screen.

This commit has to be revised when we have a proper fix upstream.

phabricator: https://phabricator.endlessm.com/T17632
bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=101520

Signed-off-by: Carlo Caione <carlo@endlessm.com>